+++ R/sits_factory.R
@@ -1,11 +1,9 @@
 #' @title Create a closure for calling functions with and without data
-#' @name .sits_function_factory
-#' @keywords internal
-#' @noRd
+#' @name sits_function_factory
 #' @author Rolf Simoes, \email{rolf.simoes@@inpe.br}
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
 #'
-#' @description This internal function implements the factory method pattern.
+#' @description This function implements the factory method pattern.
 #' Its creates a generic interface to closures in R so that the functions
 #' in the sits package can be called in two different ways:
 #' 1. Called directly, passing input data and parameters.
@@ -20,6 +18,19 @@
 #' @param data      Tibble with time series data and metadata.
 #' @param fun       Function that performs calculation on the input data.
 #' @return          A closure that encapsulates the function applied to data.
+#' @export
+sits_factory_function <- function(data, fun){
+    # if no data is given, we prepare a
+    # function to be called as a parameter of other functions
+    if (purrr::is_null(data)) {
+        result <- fun
+    } else {
+        # ...otherwise compute the result on the input data
+        result <- fun(data)
+    }
+    return(result)
+}
+
 .sits_factory_function <- function(data, fun) {
     # if no data is given, we prepare a
     # function to be called as a parameter of other functions