+++ NEWS.md
@@ -1,5 +1,16 @@
 # SITS Release History
 
+# What's new in SITS version 1.4.0
+
+### New features in SITS version 1.4.0
+* New function for image segementation based on `supercells` package
+* New version of `sits_get_data()` to extract average values of time series based on segments
+* Support for Harmonized Landsat Sentinel (HLS) collections from NASA
+* Support for probability cubes and uncertainty cubes in `sits_view()`
+* New `summary()` function to show details of data cubes and time series tibbles
+* General big fixes
+
+
 # What's new in SITS version 1.3.0
 
 ### New features in SITS version 1.3.0